app/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "mcp>=1.6.0",
    "openai-agents>=0.0.9",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
]
